requirements.txt
Flask==1.1.2
gunicorn==20.0.4
Jinja2==2.11.2
numpy==1.18.1
Werkzeug==1.0.1
Pillow==7.1.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.12.0+cpu
markupsafe==2.0.1
itsdangerous==2.0.1
flask-bootstrap
pyfluidsynth 
pretty_midi
magenta
tensorboardX